apps/api/pyproject.toml
[project]
name = "kanban-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.0",
    "asyncpg>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.14",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
